requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.0.8
django-cors-headers==3.13.0
django-taggit==3.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
